improve initialize_substate and make other changes
